Add SqlToS3Operator
